added mongodb for storing connections

+++ src/services/connections.service.ts
@@ -1,3 +1,5 @@
+import { Model } from "mongoose";
+import { ConnectionModel } from "../models";
 import { redisConnection } from "../redisConnection";
 import { UserStatusService } from "./userStatus.service";
 
@@ -5,32 +7,48 @@ export class ConnectionsService {
   private readonly redisConnection;
   private readonly userStatusService: UserStatusService;
 
-  private readonly connections: Record<number, number> = {};
+  private readonly connectionModel;
 
   constructor(userStatusService: UserStatusService) {
     this.redisConnection = redisConnection;
     this.userStatusService = userStatusService;
+    this.connectionModel = ConnectionModel;
   }
 
-  getConnectionByUser(userId: number) {
-    return this.connections[userId];
+  async getConnectionByUser(userId: number) {
+    const connection = await this.connectionModel.findOne(
+      { userId1: userId },
+      { userId2: userId }
+    );
+
+    return connection;
   }
 
   async setConnection(userId1: number, userId2: number) {
     await this.userStatusService.setUserStatus(userId1, "connected");
     await this.userStatusService.setUserStatus(userId2, "connected");
 
-    this.connections[userId1] = userId2;
-    this.connections[userId2] = userId1;
+    const newConnection = new this.connectionModel({
+      userId1: userId1,
+      userId2: userId2,
+    });
+
+    await newConnection.save();
   }
 
   async stopConnection(userId: number) {
-    const userId2 = this.connections[userId];
+    const connection = await this.getConnectionByUser(userId);
+
+    if (!connection) {
+      return;
+    }
 
     await this.userStatusService.deleteUserStatus(userId);
-    await this.userStatusService.deleteUserStatus(userId2);
+    await this.userStatusService.deleteUserStatus(connection.userId2 as number);
 
-    delete this.connections[userId2];
-    delete this.connections[userId];
+    await this.connectionModel.deleteOne({
+      userId1: userId,
+      userId2: connection.userId2,
+    });
   }
 }
